Fix an issue with authored state.
LARA was updated at some point to send the empty object `{}` as the initial authored state.
I think previously the state was null or the empty string.
This broke the table interactive since it needed a default authored state and was using
`authoredState || DEAULT_AUTHORED_STATE` to load the default state.
Initially the authoredState being `{}` meant it it would be truthy and the default state would
never be loaded.